+++ backend/core/types/src/backend-swagger.ts
@@ -205,6 +205,38 @@ export class AmiChartsService {
 
 export class ApplicationFlaggedSetsService {
   /**
+   * Meta information for application flagged sets
+   */
+  meta(
+    params: {
+      /**  */
+      page?: number
+      /**  */
+      limit?: number
+      /**  */
+      listingId: string
+      /**  */
+      view?: string
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<ApplicationFlaggedSetMeta> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/applicationFlaggedSets/meta"
+
+      const configs: IRequestConfig = getConfigs("get", "application/json", url, options)
+      configs.params = {
+        page: params["page"],
+        limit: params["limit"],
+        listingId: params["listingId"],
+        view: params["view"],
+      }
+      let data = null
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
    * List application flagged sets
    */
   list(
@@ -215,6 +247,8 @@ export class ApplicationFlaggedSetsService {
       limit?: number
       /**  */
       listingId: string
+      /**  */
+      view?: string
     } = {} as any,
     options: IRequestOptions = {}
   ): Promise<PaginatedApplicationFlaggedSet> {
@@ -226,6 +260,7 @@ export class ApplicationFlaggedSetsService {
         page: params["page"],
         limit: params["limit"],
         listingId: params["listingId"],
+        view: params["view"],
       }
       let data = null
 
@@ -271,6 +306,42 @@ export class ApplicationFlaggedSetsService {
       const configs: IRequestConfig = getConfigs("post", "application/json", url, options)
 
       let data = params.body
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Reset flagged set confirmation alert
+   */
+  resetConfirmationAlert(
+    params: {
+      /** requestBody */
+      body?: Id
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<Status> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/applicationFlaggedSets/{id}"
+
+      const configs: IRequestConfig = getConfigs("put", "application/json", url, options)
+
+      let data = params.body
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Trigger the duplicate check process
+   */
+  process(options: IRequestOptions = {}): Promise<string> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/applicationFlaggedSets/process"
+
+      const configs: IRequestConfig = getConfigs("post", "application/json", url, options)
+
+      let data = null
 
       configs.data = data
       axios(configs, resolve, reject)
@@ -1538,6 +1609,29 @@ export class MultiselectQuestionsService {
       axios(configs, resolve, reject)
     })
   }
+
+  /**
+   * Get Listings by multiselect question id
+   */
+  retrieveListings(
+    params: {
+      /**  */
+      multiselectQuestionId: string
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<Listing[]> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/multiselectQuestions/listings/{multiselectQuestionId}"
+      url = url.replace("{multiselectQuestionId}", params["multiselectQuestionId"] + "")
+
+      const configs: IRequestConfig = getConfigs("get", "application/json", url, options)
+
+      let data = null
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
 }
 
 export class ReservedCommunityTypesService {
@@ -2226,12 +2320,23 @@ export interface AmiChartUpdate {
 
   /**  */
   id?: string
+}
+
+export interface ApplicationFlaggedSetMeta {
+  /**  */
+  totalCount?: number
 
   /**  */
-  createdAt?: Date
+  totalResolvedCount?: number
 
   /**  */
-  updatedAt?: Date
+  totalPendingCount?: number
+
+  /**  */
+  totalNamePendingCount?: number
+
+  /**  */
+  totalEmailPendingCount?: number
 }
 
 export interface Address {
@@ -2521,6 +2626,9 @@ export interface Application {
   submissionType: ApplicationSubmissionType
 
   /**  */
+  reviewStatus?: ApplicationReviewStatus
+
+  /**  */
   applicant: Applicant
 
   /**  */
@@ -2637,16 +2745,22 @@ export interface ApplicationFlaggedSet {
   updatedAt: Date
 
   /**  */
-  rule: string
+  rule: EnumApplicationFlaggedSetRule
+
+  /**  */
+  ruleKey: string
 
   /**  */
   resolvedTime?: Date
 
   /**  */
-  status: EnumApplicationFlaggedSetStatus
+  listingId: string
 
   /**  */
-  listingId: string
+  showConfirmationAlert: boolean
+
+  /**  */
+  status: EnumApplicationFlaggedSetStatus
 }
 
 export interface ApplicationFlaggedSetPaginationMeta {
@@ -2683,6 +2797,14 @@ export interface ApplicationFlaggedSetResolve {
 
   /**  */
   applications: Id[]
+
+  /**  */
+  status: EnumApplicationFlaggedSetResolveStatus
+}
+
+export interface Status {
+  /**  */
+  status: string
 }
 
 export interface Asset {
@@ -2795,12 +2917,6 @@ export interface AssetUpdate {
   id?: string
 
   /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
-
-  /**  */
   fileId: string
 
   /**  */
@@ -2815,12 +2931,6 @@ export interface PaperApplicationUpdate {
   id?: string
 
   /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
-
-  /**  */
   file?: CombinedFileTypes
 }
 
@@ -2830,12 +2940,6 @@ export interface ApplicationMethodUpdate {
 
   /**  */
   id?: string
-
-  /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
 
   /**  */
   paperApplications?: PaperApplicationUpdate[]
@@ -3114,6 +3218,9 @@ export interface ApplicationCreate {
 
   /**  */
   submissionType: ApplicationSubmissionType
+
+  /**  */
+  reviewStatus?: ApplicationReviewStatus
 
   /**  */
   listing: Id
@@ -3439,13 +3546,10 @@ export interface ApplicationUpdate {
   submissionType: ApplicationSubmissionType
 
   /**  */
+  reviewStatus?: ApplicationReviewStatus
+
+  /**  */
   id?: string
-
-  /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
 
   /**  */
   deletedAt?: Date
@@ -3874,11 +3978,6 @@ export interface Email {
   appUrl?: string
 }
 
-export interface Status {
-  /**  */
-  status: string
-}
-
 export interface Confirm {
   /**  */
   token: string
@@ -3931,12 +4030,6 @@ export interface UserUpdate {
 
   /**  */
   email?: string
-
-  /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
 
   /**  */
   password?: string
@@ -4143,12 +4236,6 @@ export interface JurisdictionCreate {
 export interface JurisdictionUpdate {
   /**  */
   id?: string
-
-  /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
 
   /**  */
   name: string
@@ -4734,9 +4821,6 @@ export interface Listing {
   reviewOrderType?: ListingReviewOrder
 
   /**  */
-  listingAvailability?: ListingAvailability
-
-  /**  */
   showWaitlist: boolean
 
   /**  */
@@ -4984,6 +5068,12 @@ export interface Listing {
 
   /**  */
   closedAt?: Date
+
+  /**  */
+  afsLastRunAt?: Date
+
+  /**  */
+  lastApplicationUpdateAt?: Date
 }
 
 export interface PaginatedListing {
@@ -5165,9 +5255,6 @@ export interface ListingCreate {
 
   /**  */
   reviewOrderType?: ListingReviewOrder
-
-  /**  */
-  listingAvailability?: ListingAvailability
 
   /**  */
   applicationMethods: ApplicationMethodCreate[]
@@ -5386,6 +5473,9 @@ export interface ListingCreate {
   customMapPin?: boolean
 
   /**  */
+  lastApplicationUpdateAt?: Date
+
+  /**  */
   countyCode?: string
 
   /**  */
@@ -5401,12 +5491,6 @@ export interface ListingEventUpdate {
 
   /**  */
   id?: string
-
-  /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
 
   /**  */
   file?: AssetUpdate
@@ -5432,24 +5516,12 @@ export interface UnitAmiChartOverrideUpdate {
   id?: string
 
   /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
-
-  /**  */
   items: AmiChartItem[]
 }
 
 export interface UnitUpdate {
   /**  */
   id?: string
-
-  /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
 
   /**  */
   amiChart?: Id
@@ -5582,16 +5654,7 @@ export interface ListingUpdate {
   reviewOrderType?: ListingReviewOrder
 
   /**  */
-  listingAvailability?: ListingAvailability
-
-  /**  */
   id?: string
-
-  /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
 
   /**  */
   applicationMethods: ApplicationMethodUpdate[]
@@ -5810,6 +5873,9 @@ export interface ListingUpdate {
   customMapPin?: boolean
 
   /**  */
+  lastApplicationUpdateAt?: Date
+
+  /**  */
   countyCode?: string
 
   /**  */
@@ -5825,9 +5891,6 @@ export interface MultiselectQuestionsFilterParams {
 
   /**  */
   jurisdiction?: string
-
-  /**  */
-  applicationSection?: string
 }
 
 export interface MultiselectQuestionCreate {
@@ -5955,12 +6018,6 @@ export interface TranslationUpdate {
   id?: string
 
   /**  */
-  createdAt?: Date
-
-  /**  */
-  updatedAt?: Date
-
-  /**  */
   translations: object
 
   /**  */
@@ -6035,9 +6092,26 @@ export enum ApplicationSubmissionType {
   "paper" = "paper",
   "electronical" = "electronical",
 }
+
+export enum ApplicationReviewStatus {
+  "pending" = "pending",
+  "pendingAndValid" = "pendingAndValid",
+  "valid" = "valid",
+  "duplicate" = "duplicate",
+}
 export type AllExtraDataTypes = BooleanInput | TextInput | AddressInput
+export enum EnumApplicationFlaggedSetRule {
+  "Name and DOB" = "Name and DOB",
+  "Email" = "Email",
+}
 export enum EnumApplicationFlaggedSetStatus {
   "flagged" = "flagged",
+  "pending" = "pending",
+  "resolved" = "resolved",
+}
+export enum EnumApplicationFlaggedSetResolveStatus {
+  "flagged" = "flagged",
+  "pending" = "pending",
   "resolved" = "resolved",
 }
 export enum ApplicationMethodType {
@@ -6155,11 +6229,7 @@ export enum ListingStatus {
 export enum ListingReviewOrder {
   "lottery" = "lottery",
   "firstComeFirstServe" = "firstComeFirstServe",
-}
-
-export enum ListingAvailability {
-  "availableUnits" = "availableUnits",
-  "openWaitlist" = "openWaitlist",
+  "waitlist" = "waitlist",
 }
 
 export enum ListingEventType {