+++ sites/partners/src/listings/PaperListingForm/formTypes.ts
@@ -158,7 +158,6 @@ export const formDefaults: FormListing = {
   urlSlug: undefined,
   showWaitlist: false,
   reviewOrderType: null,
-  listingAvailability: null,
   unitsSummary: [],
   unitsSummarized: {
     unitTypes: [],