+++ sites/partners/src/listings/PaperListingForm/sections/BuildingSelectionCriteria.tsx
@@ -88,7 +88,7 @@ const LotteryResults = () => {
         content: (
           <Button
             type="button"
-            className="font-semibold uppercase text-red-700 my-0"
+            className="font-semibold uppercase text-alert my-0"
             onClick={() => {
               setCloudinaryData({
                 id: "",
@@ -136,7 +136,7 @@ const LotteryResults = () => {
             </Button>
             <Button
               type="button"
-              className="font-semibold uppercase text-red-700 my-0"
+              className="font-semibold uppercase text-alert my-0"
               onClick={() => {
                 setCloudinaryData({ ...cloudinaryData, id: "" })
                 deletePDF()
@@ -171,7 +171,7 @@ const LotteryResults = () => {
             </Button>
             <Button
               type="button"
-              className="font-semibold uppercase text-red-700"
+              className="font-semibold uppercase text-alert"
               onClick={() => {
                 setValue("buildingSelectionCriteria", "")
               }}