+++ sites/public/pages/applications/contact/name.tsx
@@ -100,7 +100,9 @@ const ApplicationName = () => {
         <Form onSubmit={handleSubmit(onSubmit, onError)}>
           <div className="form-card__group border-b" data-test-id={"application-initial-page"}>
             <fieldset>
-              <legend className="field-label--caps">
+              <legend
+                className={`text_caps-spaced ${errors.applicant?.firstName ? "text-alert" : ""}`}
+              >
                 {t("application.name.yourName")}
                 <LockIcon />
               </legend>
@@ -172,7 +174,9 @@ const ApplicationName = () => {
           </div>
 
           <div className="form-card__group">
-            <legend className="field-label--caps">
+            <legend
+              className={`text_caps-spaced ${errors.applicant?.emailAddress ? "text-alert" : ""}`}
+            >
               {t("application.name.yourEmailAddress")}
               <LockIcon />
             </legend>