+++ sites/public/pages/applications/household/member.tsx
@@ -148,7 +148,7 @@ const ApplicationMember = () => {
             <Form onSubmit={handleSubmit(onSubmit, onError)}>
               <div className="form-card__group border-b">
                 <fieldset>
-                  <legend className="field-label--caps">
+                  <legend className="text_caps-spaced">
                     {t("application.household.member.name")}
                   </legend>
 
@@ -211,10 +211,12 @@ const ApplicationMember = () => {
 
               <div className="form-card__group border-b">
                 <fieldset>
-                  <legend className="field-label--caps">
+                  <legend className="text_caps-spaced">
                     {t("application.household.member.haveSameAddress")}
                   </legend>
                   <FieldGroup
+                    fieldGroupClassName="grid grid-cols-1"
+                    fieldClassName="ml-0"
                     name="sameAddress"
                     type="radio"
                     register={register}
@@ -228,9 +230,7 @@ const ApplicationMember = () => {
 
                 {(sameAddress == "no" || (!sameAddress && member.sameAddress == "no")) && (
                   <fieldset className="mt-8">
-                    <legend className="field-label--caps">
-                      {t("application.contact.address")}
-                    </legend>
+                    <legend className="text_caps-spaced">{t("application.contact.address")}</legend>
 
                     <Field
                       id="addressStreet"
@@ -304,13 +304,15 @@ const ApplicationMember = () => {
 
               <div className="form-card__group border-b">
                 <fieldset>
-                  <legend className="field-label--caps">
+                  <legend className="text_caps-spaced">
                     {t("application.household.member.workInRegion", {
                       county: listing?.countyCode,
                     })}
                   </legend>
                   <FieldGroup
                     name="workInRegion"
+                    fieldGroupClassName="grid grid-cols-1"
+                    fieldClassName="ml-0"
                     groupNote={t("application.household.member.workInRegionNote")}
                     type="radio"
                     register={register}
@@ -324,9 +326,7 @@ const ApplicationMember = () => {
 
                 {(workInRegion == "yes" || (!workInRegion && member.workInRegion == "yes")) && (
                   <fieldset className="mt-8">
-                    <legend className="field-label--caps">
-                      {t("application.contact.address")}
-                    </legend>
+                    <legend className="text_caps-spaced">{t("application.contact.address")}</legend>
 
                     <Field
                       id="workAddress.street"
@@ -398,7 +398,7 @@ const ApplicationMember = () => {
 
               <div className="form-card__group">
                 <div className={"field " + (errors.relationship ? "error" : "")}>
-                  <label className="field-label--caps" htmlFor="relationship">
+                  <label className="text_caps-spaced" htmlFor="relationship">
                     {t("application.household.member.whatIsTheirRelationship")}
                   </label>
                   <div className="control">