+++ sites/public/pages/create-account.tsx
@@ -88,7 +88,7 @@ export default () => {
 
         <Form id="create-account" onSubmit={handleSubmit(onSubmit)}>
           <div className="form-card__group border-b">
-            <label className="field-label--caps" htmlFor="firstName">
+            <label className="text_caps-spaced" htmlFor="firstName">
               {t("application.name.yourName")}
             </label>