+++ ui-components/src/forms/Field.tsx
@@ -75,12 +75,13 @@ const Field = (props: FieldProps) => {
 
   const label = useMemo(() => {
     const labelClasses = ["label"]
-    if (props.caps) labelClasses.push("field-label--caps")
+    if (props.caps) labelClasses.push("text_caps-spaced")
     if (props.primary) labelClasses.push("text-primary")
     if (props.readerOnly) labelClasses.push("sr-only")
     if (props.type === "radio") {
       labelClasses.push("font-semibold")
     }
+    if (props.error) labelClasses.push("text-alert")
 
     return (
       <label className={labelClasses.join(" ")} htmlFor={props.id || props.name}>